Minor- Refactor classes, add new functionality to know the examples better
